// src/features/Landing/enums.ts
export enum RecoveryReasonEnum {
  Breakup = 'Breakup',
  Divorce = 'Divorce',
  Grief = 'Grief',
  Lonely = 'Lonely',
  Burnout = 'Burnout',
  Growth = 'Growth',
}

export enum TimelineEnum {
  Today = 'Today',
  Week = 'Week',
  Month = 'Month',
  Months = 'Months',
  Year = 'Year',
}

export enum FeelingEnum {
  Heartbroken = 'Heartbroken',
  Sad = 'Sad',
  Angry = 'Angry',
  Anxious = 'Anxious',
  Lonely = 'Lonely',
  Numb = 'Numb',
  Hopeful = 'Hopeful',
  Grateful = 'Grateful',
  Confident = 'Confident',
}

export enum GoalEnum {
  Overthinking = 'Overthinking',
  Sleep = 'Sleep',
  Profile = 'Profile',
  Confidence = 'Confidence',
  Heal = 'Heal',
  MoveOn = 'MoveOn',
  Motivation = 'Motivation',
}